requirements.txt
farm-haystack"[inference,crawler]"
milvus-haystack
